options: create a new file for parameters classes - Fix issue #95 - Add type checking
