+++ pyroteus/__init__.py
@@ -2,10 +2,11 @@ from pyroteus.utility import *  # noqa
 from pyroteus.time_partition import *  # noqa
 from pyroteus.recovery import *  # noqa
 from pyroteus.plot import *  # noqa
+from pyroteus.log import *  # noqa
 from pyroteus.math import *  # noqa
 from pyroteus.metric import *  # noqa
 from pyroteus.mesh_seq import *  # noqa
-from pyroteus.log import *  # noqa
+from pyroteus.options import *  # noqa
 from pyroteus.interpolation import *  # noqa
 from pyroteus.error_estimation import *  # noqa
 from pyroteus.form import *  # noqa
